requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiohttp-cors==0.7.0
aiorwlock==1.3.0
aiosignal==1.3.1
anyio==3.7.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blessed==1.20.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorful==0.5.5
deepface==0.0.80
Deprecated==1.2.14
distlib==0.3.8
fastapi==0.105.0
filelock==3.13.1
fire==0.5.0
Flask==3.0.0
flatbuffers==23.5.26
frozenlist==1.4.1
gast==0.5.4
gdown==4.7.1
google-api-core==2.15.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
gpustat==1.1.1
grpcio==1.60.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
httptools==0.6.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
keras==2.15.0
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
msgpack==1.0.7
mtcnn==0.1.1
multidict==6.0.4
numpy==1.26.2
nvidia-ml-py==12.535.133
oauthlib==3.2.2
opencensus==0.11.3
opencensus-context==0.1.3
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
platformdirs==3.11.0
prometheus-client==0.19.0
protobuf==4.23.4
psutil==5.9.6
py-spy==0.3.14
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.13
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
ray==2.8.1
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rpds-py==0.14.1
rsa==4.9
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
virtualenv==20.21.0
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
Werkzeug==3.0.1
wrapt==1.14.1
yarl==1.9.4
